fix: satisfy the clang-tidy gate and mark the gate scripts executable
The first CI run did its job. Concretely:

* the hygiene job could not run scripts/*.sh at all - the files went into the
  index without the executable bit from a Windows checkout, so every check
  exited 126 while looking like a script failure;
* enums that index arrays or name streams now state a uint8_t base;
* BoundingBox::expand_to_include says std::min/std::max rather than four ifs,
  and the crossing test names its interpolation factor instead of relying on
  operator precedence;
* the determinism assertions compare bit patterns explicitly through
  tests/support/BitPattern.hpp rather than memcmp over doubles, which is both
  what is meant and readable when it fails.

Container-conformant member type names (value_type) are exempted from the
naming check: generic code looks for exactly those spellings.

+++ core/include/paddock/core/Rng.hpp
@@ -10,7 +10,7 @@ namespace paddock::core {
 
 /// One stream per subsystem. Values are fixed and must never be renumbered:
 /// they are part of the reproducibility contract of every scenario bundle.
-enum class Subsystem : std::uint64_t {
+enum class Subsystem : std::uint8_t {
   Weather = 1,
   SoilWater = 2,
   PastureGrowth = 3,
